Create the core calculator display with CSS grid
This commit creates the basic structure of the calculator and it's
buttons. I started using flexbox but then I saw that it would probably
be interesting to do it with grid, and after some research I managed
to make it work with a pretty minimal and clean code (imo). The size of
the calculator can be adjusted just by changing the size of the buttons,
which makes it flexible.