Considerably improve calculator algorithm
Now the calculator works much more like a real one. Operations can be
chained indefinetely, showing the result every time a new operator is
clicked. After the equal sign has been pressed and the result appears
on screen,  pressing a new number resets the calculator.
The code probably has a lot of room for improvement and optimisation,
which is what I'll be working on the next commit, but so far, the
calculator seems to work pretty well.